+++ packages/react-instantsearch-hooks-dom/src/index.ts
@@ -1,6 +1,6 @@
 export * from 'react-instantsearch-hooks';
-export * from './components/Breadcrumb';
-export * from './components/Highlight';
-export * from './components/Hits';
-export * from './components/RefinementList';
-export * from './components/SearchBox';
+export * from './widgets/Breadcrumb';
+export * from './widgets/Highlight';
+export * from './widgets/Hits';
+export * from './widgets/RefinementList';
+export * from './widgets/SearchBox';
