+++ packages/react-instantsearch-hooks-dom/src/ui/lib/isModifierClick.ts
