+++ packages/react-instantsearch-hooks-dom/src/widgets/Highlight.tsx
