+++ packages/react-instantsearch-hooks-dom/src/widgets/Hits.tsx
