+++ packages/react-instantsearch-hooks-dom/src/widgets/SearchBox.tsx
