+++ packages/react-instantsearch-hooks-dom/src/widgets/__tests__/SearchBox.test.tsx
